^f Remove obsolete release helper commands (full Go and invariant validation; supporting internal cleanup)
